Renamed LogExp to LogicalExpression for clarity
