+++ src/converter/dimacs/Syntax/LogicalExpressionSyntax.ts
